Develop to main

+++ app/core/carona.py
@@ -5,6 +5,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from typing import Annotated
 from fastapi import Depends, HTTPException, status
 
+from app.database.user_carona_orm import UserCarona
 from app.utils.db_utils import get_db
 
 from app.database.carona_orm import Carona
@@ -23,12 +24,7 @@ def add_carona_to_db(
     carona_to_add: CaronaBase,
     db: Annotated[Session, Depends(get_db)]
 ) -> Carona:
-    db_carona = Carona(
-        fk_motorista = carona_to_add.fk_motorista,
-        fk_motorista_veiculo = carona_to_add.fk_motorista_veiculo,
-        hora_partida = carona_to_add.hora_partida,
-        valor=carona_to_add.valor
-    )
+    db_carona = Carona(**carona_to_add.model_dump())
     try:
         db.add(db_carona)
         db.commit()
@@ -55,12 +51,14 @@ def update_carona_in_db(
     carona_new_info: CaronaUpdate,
     db: Annotated[Session, Depends(get_db)]
 ) -> Carona:
-    if carona_new_info.fk_motorista_veiculo is not None:
-        db_carona.fk_motorista_veiculo = carona_new_info.fk_motorista_veiculo
-    if carona_new_info.hora_partida is not None:
-        db_carona.hora_partida = carona_new_info.hora_partida
-    if carona_new_info.valor is not None:
-        db_carona.valor = carona_new_info.valor
+    vagas_preenchidas = db.query(UserCarona).filter(UserCarona.fk_carona == db_carona.id).count()
+    if vagas_preenchidas > carona_new_info.vagas:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Não é possível diminuir o número de vagas disponíveis para uum número menor do que o número de vagas já preenchidas."
+        )
+    for key, value in carona_new_info.model_dump(exclude_none=True).items():
+        setattr(db_carona, key, value)
     
     try:
         db.add(db_carona)
@@ -75,10 +73,18 @@ def update_carona_in_db(
 
 def remove_carona_from_db(
     db_carona: Carona,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[Session, Depends(get_db)],
+    enforce: bool = False
 ) -> Carona:
+    vagas_preenchidas = db.query(UserCarona).filter(UserCarona.fk_carona == db_carona.id).count()
+    if vagas_preenchidas > 0 and not enforce:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Não foi possível remover a carona pois ela possui passageiros inscritos. Para removê-la, use o parâmetro 'enforce=True'."
+        )
     
     try:
+        db.query(UserCarona).filter(UserCarona.fk_carona == db_carona.id).delete()
         db.delete(db_carona)
         db.commit()
     except SQLAlchemyError as sqlae:
